Fixed errors with CORS.
For future reference it relates to how Netlify handles domain redirection. Namely, it was set to use www.danielms.site/ but the baseUrl is danielms.site

Unfortunately, this meant all assets would be referencing the wrong domain and trigger a CORS error. I made the change at Netlify and it worked instantly.