Create the plausible stats shortcode
This shortcode is used to display a timeseries and aggregate summary of a given Plausible websites montly analytics.

The shortcode usage:
    {{< plausible start=2022-01-01 end=2022-01-31 site_id=danielms.site >}}

This will render the site analytics with a Chart.js line graph and aggregate analytics data above the chart.

The site_id must match exactly with plausible's entry - case sensitive.